Update and rename ex11 to ex11.php

+++ ex11.php
@@ -3,10 +3,10 @@
 //Faça um Programa que leia um número inteiro menor que 1000 e imprima a quantidade de centenas,dezenas e unidades do mesmo, observando os termos no plural a colocação do "e", da vírgula entre
 //outros. Exemplo:
 //326 = 3 centenas, 2 dezenas e 6 unidades
-//12 = 1 dezena e 2 unidades*/
+//12 = 1 dezena e 2 unidades
 
 print "Digite um número entre 0-1000: ";
-$n=fgets(STDIN);
+$n=(int)fgets(STDIN);
 
 $qtn=$n;
